requirements.txt
Flask
Flask-Mail
Flask-SQLAlchemy
pymssql
PyMySQL
python-dotenv
SQLAlchemy==1.3.23
XlsxWriter
Werkzeug
flask-login
flask-bcrypt
wfastcgi
Flask-Cors
uuid
pandas